+++ NEWS.md
@@ -1,7 +1,7 @@
 # CFAEpiNow2Pipeline v0.2.0
 
 ## Features
-
+* Saving state exclusions to nssp-rt/state_exclusions
 * Automate tag deletion from ghcr.io
 * Editing of `SOP.md`
 * Pin r-version at 4.4.3 for CI/CD